Use decimal instead of float/double for non-integer numeric values
Replace all double/float types with decimal in protocol types, spec models,
and state records for exact representation of costs, limits, and other
user-facing numbers.

- ProtocolMessages.cs: ActionCommand.Value, ProtocolCostType.DefaultCostLimit,
  ProtocolCostValue.Value, ProtocolConstraint.Value, ProtocolCondition.Value,
  ProtocolRepeat.Value all use decimal
- RosterTypes.cs: CostState.Value uses decimal
- RosterSpecModels.cs: ActionDef.Value and ExpectedCostDef.Value use decimal
- IRosterEngine.cs: SetCostLimit parameter uses decimal
- RosterRunner.cs: simplified AssertEqual for decimal (exact equality)
- Java interop: explicit (double) casts at the boundary where decimal
  values are passed to IKVM Java methods
- NR adapter: NrCostSnapshot.Value uses decimal

Fixes #179

+++ src/BattleScribeSpec.BattleScribe/BattleScribeEngine.cs
@@ -274,10 +274,10 @@ public sealed class BattleScribeEngine : IDisposable
     /// <summary>
     /// Set cost limit for a cost type.
     /// </summary>
-    public void SetCostLimit(CostType costType, double value)
+    public void SetCostLimit(CostType costType, decimal value)
     {
         EnsureInitialized();
-        _engine.a(costType, value);
+        _engine.a(costType, (double)value);
     }
 
     /// <summary>
@@ -749,7 +749,7 @@ public sealed class BattleScribeEngine : IDisposable
     public void SetupWithPatrolAndUnit(bool withCosts = false)
     {
         var costs = withCosts
-            ? new[] { JavaModelFactory.CreateCost("pts", "pts", 100.0) }
+            ? new[] { JavaModelFactory.CreateCost("pts", "pts", 100.0m) }
             : null;
         var unitEntry = JavaModelFactory.CreateSelectionEntry("se-unit", "Marine Squad", "unit", costs: costs);
         var forceEntry = JavaModelFactory.CreateForceEntry("fe-patrol", "Patrol");
